refactor: Theme Thumbnail VO로 분리
